# server/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "WayFarer is a public bus transportation booking service.An API that will provide a fucntionality for WayFarer Admins to be able to create and manage trips.\n[GitHub Repo](https://github.com/danielotieno/way-farer-api)\n",
        "version": "1.0.0",
        "title": "WayFarer API"
    },
    "basePath": "/danielotienoke/WayFarerAPI/1.0.0",
...
    "schemes": ["https", "http"],
    "paths": {
...
        "/api/v1/trips/{tripId}": {
            "get": {
                "tags": ["Trip"],
                "summary": "Get a specific trip by ID",
                "description": "Returns a single trip",
                "operationId": "getTripById",
                "produces": ["application/json"],
                "parameters": [{
                    "name": "tripId",
                    "in": "path",
                    "description": "ID of trip to return",
                    "required": true,
                    "type": "string",
                    "format": "uuid"
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Trip"
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Trip not found"
                    }
                }
            }
        },
...
        "/api/v1/auth/signup": {
            "post": {
                "tags": ["User"],
                "summary": "Create a user",
                "description": "This can only be done by the logged in user.",
                "operationId": "createUser",
                "produces": ["application/json"],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Created user object",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/User"
                    }
                }],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "petstore_auth": {
            "type": "oauth2",
            "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
            "flow": "implicit"
        },
        "api_key": {
            "type": "apiKey",
            "name": "api_key",
            "in": "header"
        }
    },
    "definitions": {
        "User": {
            "type": "object",
            "required": ["dateCreated", "email", "firstName", "lastName", "password"],
            "properties": {
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "firstName": {
                    "type": "string",
                    "example": "Daniel"
                },
                "lastName": {
                    "type": "string",
                    "example": "Otieno"
                },
                "email": {
                    "type": "string",
                    "format": "email",
                    "example": "daniel@gmail.com"
                },
                "password": {
                    "type": "string",
                    "example": "123456789"
                },
                "dateCreated": {
                    "type": "string",
                    "format": "date"
                }
            },
            "xml": {
                "name": "User"
            }
        },
        "Trip": {
            "type": "object",
            "required": [
                "busNumber",
                "destination",
                "fare",
                "origin",
                "seatingCapacity",
                "tripDate"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "seatingCapacity": {
                    "type": "integer",
                    "example": 24
                },
                "busNumber": {
                    "type": "string",
                    "example": "RAD 276 K"
                },
                "origin": {
                    "type": "string",
                    "example": "Mombasa"
                },
                "destination": {
                    "type": "string",
                    "example": "Kigali"
                },
                "fare": {
                    "type": "number",
                    "example": 4560.0
                },
                "status": {
                    "type": "string",
                    "description": "Status of the trip",
                    "enum": ["available", "cancelled"]
                },
                "tripDate": {
                    "type": "string",
                    "format": "date"
                }
            }
        }
    }
}